Fixed case-sensitive error
*nix systems are case-sensitive with folders and
filenames, something I should know and should've
rembemered when I first build the repo on
Windows. Oh well, 'til we get some good routines
for commits things like this will happend.

Note to myself in the future, or the reader of,
this commit. Pleae take a half day and fix this
repo so that is actually functional. I'm not
very proud with the current state of the build
system.